requirements.txt
fastapi==0.109.0
pydantic==1.9.0
python-dotenv==0.19.2
uvicorn==0.15.0
SQLAlchemy==2.0.27
SQLAlchemy-Utils==0.41.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
click==8.1.7
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
ipdb==0.13.13
ipython==8.20.0
jedi==0.19.1
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic_core==2.14.6
Pygments==2.17.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.35.1
tomli==2.0.1
traitlets==5.14.1
typing_extensions==4.9.0
wcwidth==0.2.13